temp.ino
this is a script is  to checks to degrees and prints them out (this one is in fahrenheit) and this code is mostly from Adafruit so go check them out. I made the sensor read fahrenheit. and this can be worked on an arduino board and a sparkfun thing dev board or anything like that and make sure you have the Adafruit temp chip. That is the only way it will work. also install the library above (if you have not).